Option 1: Independent Subsystems
This is a proposed option as a sample of code how it could be done
differently. The code slightly rewritten to isolate the two subsystems
from each other so the program be more object oriented. Timers, cycle
counters etc. are redone too because I was just lazy to contemplate
their logic.