Wiffle Bot Light Code Beta
This is the light code that never ended up on the wiffle bot. It will
eventually be mixed with the Light Code Template as it has many
additional light functions to manipulate the lights. However there may
be a bunch of unfinished light stuff in here and I am too lazy to clean
it up.