Light Test Updates
Added some additional functionality to moveDownOnly allowing it to
simulate a rain effect. Will be ported over to other things.
Additionally left some test code in the systems just so that you know
things are working for testing. Everything was changed to a 1ms delay
allowing for very precise intentional delay using cycle speed. Made a
synchronizer that works but is really just kinda a mess.